+++ src/Raytracing.jl
@@ -23,7 +23,7 @@ export Ray, OrthogonalCamera, PerspectiveCamera, ImageTracer
 export Shape, Sphere, Plane, Torus, Vec2d, HitRecord, World
 export Pigment, UniformPigment, CheckeredPigment, ImagePigment
 export BRDF, DiffuseBRDF, SpecularBRDF, Material
-export Renderer, OnOffRenderer, FlatRenderer, PathTracer
+export Renderer, OnOffRenderer, FlatRenderer, PathTracer, PointLightRenderer
 # from Operations.jl
 export are_close, squared_norm, norm, normalize
 # from PrintFunctions.jl
@@ -38,9 +38,10 @@ export rotation_x, rotation_y, rotation_z, scaling, translation, inverse
 # from ImageTracer.jl
 export at, fire_ray, fire_all_rays!
 # from Shapes.jl
-export ray_intersection, add_shape!
+export ray_intersection, add_shape!, add_light!
 export sphere_point_to_uv, sphere_normal
 export plane_point_to_uv, plane_normal
+export is_point_visible, quick_ray_intersection
 # from Demo.jl
 export demo, demo_animation
 # from Renderers.jl